requirements.txt
albumentations==1.1.0
efficientnet_pytorch==0.7.1
Flask==2.0.2
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu
torchvision==0.11.1+cpu
